+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `DashboardMenu`: Fix `Selector` scroll on Chrome & Safari.
+
 ## [8.3.1] - 2022-02-10
 
 Fix: